Added NFT registrations
