+++ src/Blockcore.Indexer.Cirrus/Storage/Mongo/CirrusMongoBuilder.cs
@@ -27,6 +27,7 @@ namespace Blockcore.Indexer.Cirrus.Storage.Mongo
          SetDocumentMapAndIgnoreExtraElements<CirrusContractCodeTable>();
          SetDocumentMapAndIgnoreExtraElements<DaoContractComputedTable>();
          SetDocumentMapAndIgnoreExtraElements<StandardTokenComputedTable>();
+         SetDocumentMapAndIgnoreExtraElements<NonFungibleTokenComputedTable>();
 
 
          return Task.FromResult(1);
