+++ src/Blockcore.Indexer.Cirrus/Storage/Mongo/Types/NonFungibleTokenComputedTable.cs
@@ -16,13 +16,3 @@ public class NonFungibleTokenComputedTable : SmartContractComputedBase
 
    public List<string> PreviousOwners { get; set; }
 }
-
-public class Token
-{
-   public long Id { get; set; }
-   public string Address { get; set; }
-   public string Uri { get; set; }
-
-   public string OwnerAddress { get; set; }
-   public bool IsBurnet { get; set; }
-}
